rename the rule to a11y-use-next-tooltip
