+++ docs/rules/a11y-use-next-tooltip.md
