+++ src/configs/recommended.js
@@ -17,7 +17,7 @@ module.exports = {
     'primer-react/a11y-explicit-heading': 'error',
     'primer-react/no-deprecated-props': 'warn',
     'primer-react/a11y-remove-disable-tooltip': 'error',
-    'primer-react/use-next-tooltip': 'error',
+    'primer-react/a11y-use-next-tooltip': 'error',
   },
   settings: {
     github: {
