+++ src/rules/a11y-use-next-tooltip.js
